requirements.txt
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-llms-ollama>=0.2.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.5.0
pymupdf>=1.24.0
streamlit>=1.38.0
python-dateutil
